requirements.txt
# Requirements.txt file to be used with pip
tqdm
pyyaml
